backend/requirements.txt
Flask==2.2.1
pyserial==3.5
